requirements.txt
Selenium
pyinstaller
pyodbc
pandas
openpyxl
Pillow